fix attack

+++ Assets/Code/GameConfig/ScriptableObjectParser/ConfigData/Characters/CharacterAttack.cs
@@ -5,9 +5,9 @@ public readonly struct CharacterAttack
 	public int Level { get; }
 	public int ReloadAttackTime { get; }
 	public int AttackDamage { get; }
-	public float AttackCastTime { get; }
+	public int AttackCastTime { get; }
 
-	public CharacterAttack(int level, int reloadAttackTime, int attackDamage, float attackCastTime)
+	public CharacterAttack(int level, int reloadAttackTime, int attackDamage, int attackCastTime)
 	{
 		Level = level;
 		ReloadAttackTime = reloadAttackTime;
